Return to original name `DbHelper::class`.
