+++ tests/Driver/Mssql/DbHelperTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Log\Target\Db\Tests\Driver\Pgsql;
+namespace Yiisoft\Log\Target\Db\Tests\Driver\Mssql;
 
 use Throwable;
 use Yiisoft\Db\Constraint\IndexConstraint;
@@ -10,24 +10,28 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Log\Target\Db\Migration;
-use Yiisoft\Log\Target\Db\Tests\Common\AbstractMigrationTest;
-use Yiisoft\Log\Target\Db\Tests\Support\PgsqlFactory;
+use Yiisoft\Db\Schema\SchemaInterface;
+use Yiisoft\Log\Target\Db\DbHelper;
+use Yiisoft\Log\Target\Db\Tests\Common\AbstractDbHelperTest;
+use Yiisoft\Log\Target\Db\Tests\Support\MssqlFactory;
 
 /**
- * @group pgsql
+ * @group Mssql
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class MigrationTest extends AbstractMigrationTest
+final class DbHelperTest extends AbstractDbHelperTest
 {
+    protected string $logTime = SchemaInterface::TYPE_DATETIME;
+    protected string $messageType = SchemaInterface::TYPE_STRING;
+
     protected function setUp(): void
     {
         // create connection dbms-specific
-        $this->db = (new PgsqlFactory())->createConnection();
+        $this->db = (new MssqlFactory())->createConnection();
 
         // set table prefix
-        $this->db->setTablePrefix('pgsql_');
+        $this->db->setTablePrefix('mssql_');
 
         parent::setUp();
     }
@@ -43,7 +47,7 @@ final class MigrationTest extends AbstractMigrationTest
      */
     public function testVerifyTableIndexes(string $tableWithPrefix, string $table): void
     {
-        Migration::ensureTable($this->db, $tableWithPrefix);
+        DbHelper::ensureTable($this->db, $tableWithPrefix);
 
         $schema = $this->db->getSchema();
         $table = $this->db->getTablePrefix() . $table;
