--- tests/Driver/Mssql/MigrationTest.php
@@ -1,78 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Log\Target\Db\Tests\Driver\Mssql;
-
-use Throwable;
-use Yiisoft\Db\Constraint\IndexConstraint;
-use Yiisoft\Db\Exception\Exception;
-use Yiisoft\Db\Exception\InvalidArgumentException;
-use Yiisoft\Db\Exception\InvalidConfigException;
-use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Schema\SchemaInterface;
-use Yiisoft\Log\Target\Db\Migration;
-use Yiisoft\Log\Target\Db\Tests\Common\AbstractMigrationTest;
-use Yiisoft\Log\Target\Db\Tests\Support\MssqlFactory;
-
-/**
- * @group Mssql
- *
- * @psalm-suppress PropertyNotSetInConstructor
- */
-final class MigrationTest extends AbstractMigrationTest
-{
-    protected string $logTime = SchemaInterface::TYPE_DATETIME;
-    protected string $messageType = SchemaInterface::TYPE_STRING;
-
-    protected function setUp(): void
-    {
-        // create connection dbms-specific
-        $this->db = (new MssqlFactory())->createConnection();
-
-        // set table prefix
-        $this->db->setTablePrefix('mssql_');
-
-        parent::setUp();
-    }
-
-    /**
-     * @dataProvider tableListIndexesProvider
-     *
-     * @throws Exception
-     * @throws InvalidArgumentException
-     * @throws InvalidConfigException
-     * @throws NotSupportedException
-     * @throws Throwable
-     */
-    public function testVerifyTableIndexes(string $tableWithPrefix, string $table): void
-    {
-        Migration::ensureTable($this->db, $tableWithPrefix);
-
-        $schema = $this->db->getSchema();
-        $table = $this->db->getTablePrefix() . $table;
-
-        /** @psalm-var IndexConstraint[] $indexes */
-        $indexes = $schema->getTableIndexes($tableWithPrefix, true);
-
-        $this->assertSame(['id'], $indexes[0]->getColumnNames());
-        $this->assertSame("PK_$table", $indexes[0]->getName());
-        $this->assertTrue($indexes[0]->isUnique());
-        $this->assertTrue($indexes[0]->isPrimary());
-
-        $this->assertSame(['category'], $indexes[1]->getColumnNames());
-        $this->assertSame("IDX_$table-category", $indexes[1]->getName());
-        $this->assertFalse($indexes[1]->isUnique());
-        $this->assertFalse($indexes[1]->isPrimary());
-
-        $this->assertSame(['level'], $indexes[2]->getColumnNames());
-        $this->assertSame("IDX_$table-level", $indexes[2]->getName());
-        $this->assertFalse($indexes[2]->isUnique());
-        $this->assertFalse($indexes[2]->isPrimary());
-
-        $this->assertSame(['log_time'], $indexes[3]->getColumnNames());
-        $this->assertSame("IDX_$table-time", $indexes[3]->getName());
-        $this->assertFalse($indexes[3]->isUnique());
-        $this->assertFalse($indexes[3]->isPrimary());
-    }
-}
