+++ tests/Driver/Sqlite/DbHelperTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Log\Target\Db\Tests\Driver\Oracle;
+namespace Yiisoft\Log\Target\Db\Tests\Driver\Sqlite;
 
 use Throwable;
 use Yiisoft\Db\Constraint\IndexConstraint;
@@ -11,26 +11,26 @@ use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Schema\SchemaInterface;
-use Yiisoft\Log\Target\Db\Migration;
-use Yiisoft\Log\Target\Db\Tests\Common\AbstractMigrationTest;
-use Yiisoft\Log\Target\Db\Tests\Support\OracleFactory;
+use Yiisoft\Log\Target\Db\DbHelper;
+use Yiisoft\Log\Target\Db\Tests\Common\AbstractDbHelperTest;
+use Yiisoft\Log\Target\Db\Tests\Support\SqliteFactory;
 
 /**
- * @group Oracle
+ * @group sqlite
  *
  * @psalm-suppress PropertyNotSetInConstructor
  */
-final class MigrationTest extends AbstractMigrationTest
+final class DbHelperTest extends AbstractDbHelperTest
 {
     protected string $idType = SchemaInterface::TYPE_INTEGER;
 
     protected function setUp(): void
     {
         // create connection dbms-specific
-        $this->db = (new OracleFactory())->createConnection();
+        $this->db = (new SqliteFactory())->createConnection();
 
         // set table prefix
-        $this->db->setTablePrefix('oci_');
+        $this->db->setTablePrefix('sqlite3_');
 
         parent::setUp();
     }
@@ -46,7 +46,7 @@ final class MigrationTest extends AbstractMigrationTest
      */
     public function testVerifyTableIndexes(string $tableWithPrefix, string $table): void
     {
-        Migration::ensureTable($this->db, $tableWithPrefix);
+        DbHelper::ensureTable($this->db, $tableWithPrefix);
 
         $schema = $this->db->getSchema();
         $table = $this->db->getTablePrefix() . $table;
@@ -54,24 +54,19 @@ final class MigrationTest extends AbstractMigrationTest
         /** @psalm-var IndexConstraint[] $indexes */
         $indexes = $schema->getTableIndexes($tableWithPrefix, true);
 
-        $this->assertSame(['category'], $indexes[0]->getColumnNames());
-        $this->assertSame("IDX_$table-category", $indexes[0]->getName());
+        $this->assertSame(['log_time'], $indexes[0]->getColumnNames());
+        $this->assertSame("IDX_$table-time", $indexes[0]->getName());
         $this->assertFalse($indexes[0]->isUnique());
         $this->assertFalse($indexes[0]->isPrimary());
 
-        $this->assertSame(['id'], $indexes[1]->getColumnNames());
-        $this->assertSame("PK_$table", $indexes[1]->getName());
-        $this->assertTrue($indexes[1]->isUnique());
-        $this->assertTrue($indexes[1]->isPrimary());
+        $this->assertSame(['level'], $indexes[1]->getColumnNames());
+        $this->assertSame("IDX_$table-level", $indexes[1]->getName());
+        $this->assertFalse($indexes[1]->isUnique());
+        $this->assertFalse($indexes[1]->isPrimary());
 
-        $this->assertSame(['log_time'], $indexes[2]->getColumnNames());
-        $this->assertSame("IDX_$table-time", $indexes[2]->getName());
+        $this->assertSame(['category'], $indexes[2]->getColumnNames());
+        $this->assertSame("IDX_$table-category", $indexes[2]->getName());
         $this->assertFalse($indexes[2]->isUnique());
         $this->assertFalse($indexes[2]->isPrimary());
-
-        $this->assertSame(['level'], $indexes[3]->getColumnNames());
-        $this->assertSame("IDX_$table-level", $indexes[3]->getName());
-        $this->assertFalse($indexes[3]->isUnique());
-        $this->assertFalse($indexes[3]->isPrimary());
     }
 }
